CASMCMS-7699: Read and report the BSS_REFERRAL_TOKEN on node boot
Change the BOS state reporter to look for the BSS_REFERRAL_TOKEN in the
kernel command line parameters. BSS returns this token when boot
artifacts are associated with nodes. BOS receives this token after it
performs such an association. The BOS state reporter reports this token
when the node boots. Thus, BOS, which associated the boot artifacts it
told BSS about with this token, can tell what the boot artifacts are on
the node based on its token-to-artifact map.