Copy V1 session templates from etcd to V2's redis database
BOS V2 stores its session templates in a redis database. This mod
copies the existing V1 SessionTemplates from their storage in etcd to
redis. It copies them, but does not delete them from etcd. This
preserves the ability to rollback to the V1 release and retain access to
the session templates.

It contains a migration script that is run as a Helm hook on upgrade.
This migration script migrates BOS V1 session templates from the
V1 etcd database to the V2 redis database.

The V1 session templates are pruned, so that they only contain
content appropriate for a V2 session template.

Any V2 session templates that exist will not be overwritten during
migration. This allows migration to be done repeatedly without
risk of losing a V2 session template.

Because the migration job reads directly from etcd, it must wait
for those pods to be up before it runs.