Copy V1 session templates from etcd to V2's redis database
BOS V2 stores its session templates in a redis database. This mod
copies the existing V1 SessionTemplates from their storage in etcd to
redis. It copies them, but does not delete them from etcd. This
preserves the ability to rollback to the V1 release and retain access to
the session templates.