CASMCMS-6169: Add new Discovery Agent and options
This mod adds a new discovery agent that periodically aggregates
information from HSM about nodes that may be booted. We do not filter
the provided xnames here, even if we do not intend to boot them (yet).
We also do not remove any xname entry that goes missing, as there may be
desire to maintain records should the xname re-appear.

This mod additionally adds a new polling frequency unique to the
discovery agent to decrease the overall impact on HSM state management
for events that do not occur with great frequency (newly added
hardware).