CASMCMS-7290: Process the BSS_REFERRAL_TOKEN
Upon an update (PUT or PATCH) to BSS, BSS will return a
BSS_REFERRAL_TOKEN. This token identifies the boot artifacts (kernel,
kernel parameters, and initrd). BSS will pass this token down to the
nodes via their kernel commandline parameters. The BOS client on the
nodes reports this token back up to the BOS API. In this way, BOS
knows which boot artifacts the node booted with.

This mod processes the BSS_REFERRAL_TOKEN. It maps the returned token to
the boot artifacts for later reference. It updates each nodes' desired
state with the token. This token is used during comparisons between the
desired and actual state.