CASMCMS-7290: Process the BSS_REFERRAL_TOKEN
Upon an update (PUT or PATCH) to BSS, BSS will return a
BSS_REFERRAL_TOKEN. This token identifies the boot artifacts (kernel,
kernel parameters, and initrd). BSS will pass this token down to the
nodes via their kernel commandline parameters. The BOS client on the
nodes reports this token back up to the BOS API. In this way, BOS
knows which boot artifacts the node booted with.

This mod processes the BSS_REFERRAL_TOKEN. It maps the returned token to
the boot artifacts for later reference. It updates each nodes' desired
state with the token.

Add the name to the API under the V2ComponentActualState.

When a compute node reports its BSS token for its actual state, look up
the boot artifacts that correspond to that token. Populate the node's
actual state with those artifacts.

There is a work-around using the excludeOutboundPort to allow the
power_on operator to
communicate with the redis database. Without this work-around, the
operator receives a 104, connection reset by peer error when it attempts
to talk with the database.

Unrelated fixes:

The CFS get_components method now returns actual components. Before it
returned nothing.

Fix DesiredConfigurationSetInCFS. It needs to query CFS using a string
of xnames as the input parameter for 'ids', not a detailed dictionary
of the components.