CASMCMS-7290: Comment out record_boot_artifacts
The record_boot_artifacts call was causing problems connecting to the
database. I'm commenting it out. We need to sort out whether the
table mapping BSS tokens to boot artifacts should be exposed via the
API or not. If it is, then we'll go through the API to update the table
rather than writing straight to the database.