CASMCMS-7290: Copy dbs package underneath server.bos package.
Hack. This is a hack to test out the boot_artifacts access via the
components endpoint. Our BOS hierarchy is wrong in that it has two 'bos'
packages at different levels in the hierarchy. It should only have one.
I copied the dbs package underneath server.bos, so it was accessible
by the modules located there. This will need to be undone later when
we reorganize the code to moven server underneath the src.bos package.