CASMCMS-7794: Build BOS rpms for SLES15SP2 and SLES15SP3 distributions.
Build the BOS rpms for the SLES15SP2 and SLES15SP3 distributions.
The COS compute node images are SLES-based, so we need to build the
rpm with the distribution environment.

To prevent the accumulation of files from previous stages in the build,
clean up the contents of the RPM build directories after publishing its
contents. This does clean-up ahead of time, so that SLES15SP2 files are not
included with SLES15SP3 files.

Publish the Helm chart and Docker image ahead of the RPMs.
This helps speed up development because developers are less interested in
the RPM results.

(cherry picked from commit cd57456cadf7c3a9173e4bc585f592ff6ef11d27)