CASMCMS-7845: Verify session creation
Ensure that the Session template being referenced exists, and that it
contains boot artifacts if the operation is a boot or reboot.

Provide environment variables for S3 access. These were not a part of
the environment variables for the BOS API. They were only present for
BOA.

When BOS cannot read the manifest file, different exceptions can be
thrown. Rather than allowing unexpected exceptions to be raised and not
be caught, I opted to catch them all -- i.e. catch Exceptions --
and raise a recognizable error.