CASMCMS-7846: Create a sessiontemplatevalidate endpoint
The sessiontemplatevalidate endpoint validates an existing Session
Template. This allows users to create incomplete or incorrect Session
Templates and later validate them.

It will validate a session template to see that it has at least one boot set.
The boot sets should specify their targets.
Boot artifacts should be present, but only the kernel is required.
The initrd and kernel boot parameters are not required for validation.
However, in most cases they are needed for a successful boot.
The absence of initrd and kernel parameters in a session
template cause warnings. Only a missing kernel causes an error.

This moves the validate_boot_sets method into its own module.
It is utilized both at Session template creation and Session creation.

Do not immediately exit if one of the boot artifacts is missing when
creating a S3BootImageMetaData instance. This prevents us from throwing
errors when something like rootfs is missing.