CASMCMS-7846: Use correct session template name
Use the name of the session template during its validation when issuing
warnings. During session creation, the name is taken from an existing
session template. During session template creation, the name is taken
from the endpoint used to create the session template, which will end
up being the template's name. But during session template creation,
the name in the supplied payload is not used because it will be
overwritten by the name used as part of the request, basically the
name affixed to the endpoint.