API Side Query and Deletion
This mod changes the exposed option from an integer (in seconds) to a
string, which is expressed as a period of time using 'h', 'm', 'w' and
'd', e.g. '7d' is 7 days time.

It removes operator side filtering for age based on session stop time,
to session start time filtering on the API side.

This simplifies the operator code base and allows future granularity in
seconds, should we find the need.