CASMCMS-7846:
Do not immediately exit if one of the boot artifacts is missing when
creating a S3BootImageMetaData instance. This prevents us from throwing
errors when something like rootfs is missing.

Downgraded the absence of initrd and kernel parameters in a session
template to warnings. Only a missing kernel causes an error.

Improved the error messages.