CASMCMS-7940: Add network policy for power-on operator to talk to DB
Add a network policy to allow the power-on operator to talk with the
cray-bos-db. The power-on operator talks with the database because it
writes to the table mapping BSS tokens to boot artifacts. This data is
not exposed on an API endpoint because it was deemed uninteresting to
users. Therefore, rather than writing to an API endpoint, the operator
talks straight to the database.

Locking down, so it only admits traffic on the redis port.