CASMCMS-7948: Add a cfs_components_dict attribute to DesiredConfigurationSetInCFS
Pass the CFS components to the _match method for
DesiredConfigurationSetInCFS filter using either an optional input parameter
or a class attribute rather than. Making the input parameter options allows
the DesiredConfigurationSetInCFS filter to conform to the other filters whose
_match methods only take a single components parameter. This allows it to be
usable with the NOT filter. However, the status operator does not call the
DesiredConfigurationSetInCFS filter's _filter method, so it cannot rely on
the _filter method setting up the cfs_components_dict inside the Class
instance. Therefore, the _match needs the optional cfs_componet parameter,
so that the status operator can pass it in directly.

Include the negated filter's _filter method in the resultant NOT
instance. The _filter may have effects that need to be implemented.