CASMCMS-7893: Add a SessionsStatusEndpoint to the BOS client.
Add a SessionsStatusEndpoint to the BOS client. This class provides
access to the status for a BOS session. You can create it, update it,
get it, and delete it. It acts in a similar fashion to the other
endpoints in BosClient except that the session ID is part of the URL.
This difference necessitated writing a new class that didn't inherit
from the BaseBosEndpoint class.