CASMCMS-8001: Provide session data when operator updates desired state
Provide session data when updating the component's desired state. This
is a hacky way of proving that you are an operator. If the session data
is not provided, then the assumption is that it was a user and not an
operator making the change. When users change a component's desired
state, we decided that change means the component should no longer be
part of a session and therefore the component's session is blanked.