CASMCMS-8025: Properly compare actual and desired kernel parameters
To properly compare kernel boot parameters between BOS sessions, the
parameters that can change dynamically must be stripped away first.
Thus, the spire_join_token needs to go. Additionally, the bos_session_id
needs to go. BOS V1 provided the bos_session_id on the kernel boot
parameters, but this mod is removing it because it did not provide
a lot of value in the first place.