CASMCMS-8023: Correct the way BOS Status Reporter Sleeps
When a non-default value is passed in to bos-status-reporter through
/proc/cmdline, it is first parsed as a datetime.timedelta object. On
computes, BOS automatically injects this value onto /proc/cmdline. When
this happens, BOS assumes that a non-default value is to be used, so it
needs to convert this default (4h) into a period of time, in seconds,
that are required to sleep before reporting.

Unfortunately, a multiplicative set operation was being used in addition
to a conversion from datetime.timedelta to total_seconds, which
increased the overall reporting window by an additional multiplicative
factor. As a result, BOS-status-reporter was sleeping way too long to
report overall boot artifacts when this value was specified.

This mod corrects the problem and adds additional logging information to
report the actual value being used by bos-status-reporter.