CASMCMS-8050: Create POST method delete others for session status
Delete the DELETE, PUT, and PATCH methods for the session status. They
were never needed and were only created as part of a misunderstanding.
However, the POST method is needed, and it is used to inform BOS V2 to
save the session status to the database. As a result, the POST method
does not need any data.