CASMCMS-8101: Migration job should talk only to new cray-bos pods
The migration Helm hook needs to only talk with the newly spun up
pods. It should not talk with the pods that are spinning down
because their V1 endpoints talk with the old etcd database. We need
the V1 endpoints to be talking with the new redis database, so that
the session templates are migrated to the redis database. Talking
to the etcd DB means the session templates will be stored there, but
will be inaccessible because the new endpoints ignore etcd. In effect,
upon upgrade the session templates look like they disappeared instead
of migrating to the new database.

To ensure it is talking to the new pods, the Helm hook job needs to
identify them by their labels. It then gets one of the new pods'
IPs and talks only to it.

Side note, the migration script needs to wait for Istio to be
running before it starts otherwise it hits its retry limit
futilely attempting to make connections. Fortunately, there is an
Istio pod annotation that handles this situation.

Changes:
* Increase the version of the base chart cray-service that BOS is
dependent upon. Do this because we want the additional pod labels
functionality introduced through the cray-service mod that
addressed CASMPET-5895.
** Note: This is a major version increase.
* The App Version used to label the pods is the Docker version,
which is calculated using gitversion, so the Docker version
correlates to the released application version.
* Add in App Version label to the cray-bos API pods and the
operators.
* Add the App Version to the environment variables for the migration
job.
* Add pods to the list of resources that the migration job can query.
* Add an http port to the migration job pod
** Pass the container's port number to the migration job container via an
enviroment variable. With this, the migration job can talk directly to
the cray-bos container within the pod.
* Correct the v1 to v2 endpoint.  Was using the v1 endpoint for the
v2 session template creation.