CASMCMS-8158: Allow users to globally influence lifecycle of staged data
Previously, staged data was always moved from the staged area of a
component to the desired state; at that time, staged data was zero'd. As
a result, staging was viewed as a one-time option.

The introduction of this mod updates the default behavior to keep staged
information present in the component record, so that subsequent calls
can move this staged information back onto the component time and time
again.

Update the option to be named 'clear_stage'

Update openapi.yaml option name.

Correct import location of library
Add call into function routine.