CASMCMS-8754: Make BOS V2 status operator resilient to power errors
When CAPMC returns errors, handle them and disable the associated nodes
instead of panicking and halting all forward progress.

Added some new classes.

Handle troublesome CAPMC errors that don't point fingers and indentify
which nodes experienced the errors.

If CAPMC returns an error that cannot be associated with an individual
node, then reissue the CAPMC command for each individual component.
Under that condition, Any components returning an error is guaranteed
to be the cause of the error, so they are associated with the error
and disabled in order to not cause BOS to attempt to retry them.

As a consequence of handling CAPMC-returned errors, the power
operators need to be able to disable nodes. When they receive
errors from CAPMC, they can declare those nodes disabled.

Create a power_operator_base class because error handling is common
to all three power operators. This is an abstract base class
that each of the power operators inherits from. It collects all of
the error handling into one place, so that it is not spread across
the three power operators.