CASMCMS-8754: Handle troublesome CAPMC errors that don't point fingers
If CAPMC returns an error that cannot be associated with an individual
node, then reissue the CAPMC command for each individual component.
Any components returning errors are guaranteed to be the cause of the
error, so they are associated with the error and disabled in order to
not cause BOS to attempt to retry them.