CASMCMS-8754: Fix up how I am creating a list
Using the list(var) construct, Python creates a list out of the contents
of the variable var. If var is a string, each character in the string
is turned into an individual list element. However, explicitly creating
a list like this [var] creates a list where the contents of var is a
single element in the list. This latter is what I had intended, but I
mistakenly used the list(var) construct.