CASMCMS-8830: Fix return type error
When there are no nodes in a session due to them being eliminated for
one reason or another, ensure that hsm.py's get_components method
returns the expected type of object.

In the case that there are no nodes within a session to act upon,
raise an exception, so that the session is marked as having failed and
has its error attribute filled in appropriately.

(cherry picked from commit ffd36b5791bcac5806e9591d7efad709a359759b)