CASMCMS-8835: Do not prematurely filter out disabled nodes
To truly allow the 'include_disabled' nodes capability, the power
operators cannot filter out the disabled nodes. This mod removes that
filter. The session's 'include_disabled' nodes attribute will be used
in the session setup operator to decide whether the nodes will be
enabled in BOS or not. If disabled nodes are not included, then they
will never be enabled and never acted upon. Thus, including them in
the query to HSM will not hurt anything because they will never be
acted upon.

When filtering nodes by architecture, do not also filter by whether the
nodes are enabled in HSM. This is done at a later step in the process.
Doing this earlier thwarts the later filtering.