CASMCMS-8879: Unpin the Alpine version from 3.18 back to 3
The python virtual environment change should enable us to move ahead
to later versions of Alpine 3 beyond 3.18. These later versions
should not cause problems due PEP-668 because the resolution to
CASMCMS-8879 fixed that issue. Pinning the version to 3.18
prevented us from encountering the PEP-668 problem because it
pinned to a version of Alpine that did not introduce the issue.