CASMCMS-9081: Created an explicit authorized_requests_retry_sessions
The original requests_retry_sessions function that is specific to the
bos-reporter code looked nearly identical to the
requests_retry_sessions function in bos.common.utils; however, it
added in an authorization token. This subtlety was easily overlooked
and led to the removal of this code and replacement with the code
from bos.common.utils. This led to a failure due to the lack of
authorization token.

Thus, authorized_requests_retry_sessions was created to both utilize
the requests_retry_sessions code in bos.common.utils and explicitly
call out the difference via the name.