CASMCMS-9078: BOS tags SBPS images
The Scalable Boot Provisioning Service (SBPS) provides root
filesystems to nodes when they boot. The images containing
these root filesystems need to be tagged in the Image
Management Service with 'sbps-project: true' before SBPS projects
them to the nodes. With this mod, BOS tags any rootfs it is booting
a node with with this key/value tag to ensure that it is projected
during booting. This relieves the admin from needing to manually
tag the image. This is a quality of life mod.